[KAIZEN-0] splittet redux-oppsett i logiske filer
